fix(docs): harden navigation evaluation boundaries
